server/requirements.txt
pillow_heif
Pillow
flask